New command structure. Removed init functions
