+++ cmd/pyroscope/command/banner.go
@@ -1,4 +1,4 @@
-package cmd
+package command
 
 import (
 	"strings"
