+++ cmd/pyroscope/command/usage.go
@@ -1,4 +1,4 @@
-package cmd
+package command
 
 import (
 	"fmt"
