--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/pyroscope-io/pyroscope/cmd"
-
-func main() {
-	cmd.Execute()
-}
