+++ pyroscope.go
@@ -1,0 +1,4 @@
+package pyroscope
+
+// this file is only here because otherwise pkger fails.
+// See: https://github.com/markbates/pkger/issues/132
